Moved from CSrankings provided aliases to DBLP parsed aliases.
CSrankings only provides alias information for university academics,
not for industry researchers. Instead of using the directly
available aliasing information we now parse our own as part of the
`www` parsing from DBLP, splitting out alias information when we
do the initial pass of the SAX stream.

This allows us to go from 1841 replaced aliases in today's dataset
to 4752, great improvement for better data. We may have to double
check if there are any flaws in the DBLP data or if CSrankings
did any manual improvement.